feat: Add cache_clear method to AdGuardHome class
This commit adds a new method `cache_clear` to the `AdGuardHome` class in the `adguardhome.py` file. The `cache_clear` method allows clearing the cache of the AdGuard Home instance. It sends a POST request to the `/control/cache_clear` endpoint.

This change was made to provide a convenient way for users to clear the cache of the AdGuard Home instance.